\name{ME_ECharmserve}
\alias{ME_ECharmserve}
\title{
Harmonization models for soil electrical conductivity
}
\description{
Mixed effects models for harmonizing electrical conductivity (EC)
}
\details{
Suit of mixed-effects models
}
\note{
Internet connectivity is needed for the function to work.
}
