\name{soil}
\alias{soil}
\docType{data}
\title{
Sample soil dataset for salinity mapping
}
\description{
Horizon sample dataset for mapping soil salinity
}
\usage{data("soil")}
\format{
  The format is:
Formal class 'SpatialPointsDataFrame' [package "sp"] with 5 slots
  ..@ data       :'data.frame':	152 obs. of  17 variables:
  .. ..$ Sample   : Factor w/ 152 levels "1","10","100",..: 1 65 76 87 98 109 120 131 142 2 ...
  .. ..$ ProfileID: Factor w/ 87 levels "1","2","3","4",..: 5 53 53 55 55 56 6 7 57 8 ...
  .. ..$ Latitude : num [1:152] -30.2 -30.3 -30.3 -30.3 -30.3 ...
  .. ..$ Longitude: num [1:152] 62.2 62.1 62.1 62.1 62.1 ...
  .. ..$ Horizon  : Factor w/ 2 levels "A","B": 1 1 2 1 2 2 1 1 2 1 ...
  .. ..$ Depth    : Factor w/ 43 levels "0 - 100","0 - 17",..: 8 14 37 8 29 42 8 8 38 8 ...
  .. ..$ Sand     : num [1:152] 43.2 61.2 57.2 55.2 65.2 83.2 63.2 63.2 45.2 59.2 ...
  .. ..$ Silt     : num [1:152] 44 24 29 32 22 9 24 24 40 24 ...
  .. ..$ Clay     : num [1:152] 12.8 14.8 13.8 12.8 12.8 7.8 12.8 12.8 14.8 16.8 ...
  .. ..$ OC       : num [1:152] 0.36 0.465 0.39 0.36 0.42 0.87 0.075 0.375 0.84 0.33 ...
  .. ..$ PH       : num [1:152] 8.6 8.37 8.31 8.76 7.81 ...
  .. ..$ EC       : num [1:152] 0.8 2.58 0.98 0.532 1.87 18.5 0.43 0.302 0.345 2.7 ...
  .. ..$ CaCo3    : num [1:152] 15.2 18.5 20.5 15.8 20 ...
  .. ..$ K        : num [1:152] 67 162 120 124 177 91 127 72 123 158 ...
  .. ..$ Na       : num [1:152] 1073 707 689 646 691 ...
  .. ..$ CEC      : num [1:152] 6 11 18 9 10.4 6 6.4 16 10 4.9 ...
  .. ..$ ESP      : Factor w/ 22 levels "0","1","10","11",..: 11 19 17 20 20 8 17 20 15 11 ...
  ..@ coords.nrs : num(0)
  ..@ coords     : num [1:152, 1:2] 420924 418226 418226 415334 415334 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : NULL
  .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  ..@ bbox       : num [1:2, 1:2] 386582 3343117 427796 3386711
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:2] "coords.x1" "coords.x2"
  .. .. ..$ : chr [1:2] "min" "max"
  ..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
  .. .. ..@ projargs: chr "+proj=utm +zone=41 +datum=WGS84 +units=m +no_defs"
}
\details{
A dataset with 87 points of soil horizons for mapping salinity
}
\source{
Hypothetical dataset for salinity mapping
}
\references{
Hypothetical dataset for salinity mapping
}
\examples{
data(soil)
str(soil)
}
\keyword{datasets}
